+++ src/main/java/com/matchday/matchdayserver/match/model/dto/response/MatchMemoResponse.java
@@ -1,9 +1,11 @@
-package com.matchday.matchdayserver.matchmemo.model.dto;
+package com.matchday.matchdayserver.match.model.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
+@Setter
 @Builder
 public class MatchMemoResponse {
   private Long matchId;
